aether/cloud-bridge/requirements.txt
pymavlink>=2.4.41
awsiotsdk>=1.21.4
paho-mqtt>=1.6.1
